+++ Archive/Alpha/Alpha 0.2 /Irrelevant modules/Pre_Comp_Prep.py
@@ -1,4 +1,3 @@
-import cv2 as cv
 import numpy as np
 
 
@@ -7,103 +6,64 @@ def Pre_comp_prep(crop_dim,bl_wh_lvls):     # Case by case comparison
     def preliminary_creation_of_arrays():
         def contour_constructor():
             # Start of Contour constructor
-
             # White Keys
-            split_fac_white = crop_dim[1]/52
+            split_fac = crop_dim[1]/52
             y_white = bl_wh_lvls[1]
             white_k_contours = []
             contour_temp = []
-            cnt = 0
+            cnt = 0 # used for case selection, takes values = [0,1,2]
             for i in range(53):
                 if cnt == 0:    # For first white key only  # Left Bound
-                    i *= split_fac_white 
-                    print(i)
+                    i *= split_fac 
                     l_b, l_t = (i,y_white), (i,0)
-
                     contour_temp.append(l_b)    # LB_vortex  [0]
                     contour_temp.append(l_t)    # LT_vortex  [1]
                     cnt += 1
-
                 elif cnt == 1:  # For first white key only    # Right Bound  and calculate other
-                    i *= split_fac_white 
-                    print(i)
-
+                    i *= split_fac
                     r_t, r_b = (i,0), (i,y_white)   # Inverse order to fullfill Clockwise order of contour vertices
-
                     contour_temp.append(r_t)    # RT_vortex  [2]
                     contour_temp.append(r_b)    # RB_vortex  [3]
-                    
                     # Commit Contour
                     white_k_contours.append(contour_temp)
                     cnt += 1
-
                 else:   # For all other white Keys
                     l_b,l_t = contour_temp[3],contour_temp[2] # Grab right bounds of previous key # Inverted fullfill Clockwise order of contour vertices
                     contour_temp = []   # Empty temp. array
-
-                    i *= split_fac_white 
-                    print(i)
-
+                    i *= split_fac 
                     r_t, r_b = (i,0), (i,y_white)
-
                     contour_temp = [l_b,l_t,r_t,r_b]    # Join
                     white_k_contours.append(contour_temp)   #   Commit
 
-
-
-
             # Black Keys
-            split_fac = crop_dim[1]/89
+            split_fac = crop_dim[1]/89  # New split fact
             y_black = bl_wh_lvls[0]
-            skip_list = [0,3,8,15,20,27,32,39,44,51,56,63,68,75,80,87]
-            cnt = 0
+            skip_list = [0,3,8,15,20,27,32,39,44,51,56,63,68,75,80,87] # Works but there must be better "more intelligent method"
+            mode_flag = True
             black_k_contours = []
-            contour_temp = [] 
+            contour_temp = []   # Reset temp array
             for i in range(88):
                 if i in skip_list:
                     continue
-
-                else:   # lft lat
-                    if cnt == 0:
+                else:   # [lft_bot, lft_top]
+                    if mode_flag == True:    # First case handle
                         i *= split_fac
-                        xy1,xy2 = (i,0),(i,y_black)
-                        # cv.line(out, xy1, xy2, (255,0,0), thickness=1)  # Left
+                        l_b, l_t = (i,y_black), (i,0)
+                        contour_temp.append(l_b)    
+                        contour_temp.append(l_t)
+                        mode_flag = False
 
-                        contour_temp.append(xy1)
-                        contour_temp.append(xy2)
-                        cnt += 1
-
-                    else:   # rgt lat + Top + Bot
+                    else:   # [rgt_top, rgt_bot]
                         i *= split_fac
-                        xy1,xy2 = (i,0),(i,y_black)
-                        # cv.line(out, xy1, xy2, (255,0,0), thickness=1)  # Rgt
-
-                        top1, top2 = contour_temp[0],xy1
-                        # cv.line(out, top1, top2, (255,0,0), thickness=1)  # Top
-
-                        bot1, bot2 = contour_temp[1],xy2
-                        # cv.line(out, bot1, bot2, (255,0,0), thickness=1)  # Bot
-
-                        contour_temp.reverse()  # For Clockwise order requirement
-
-                        # Commit
-                        # Top
-                        # contour_temp.append(top1)
-                        contour_temp.append(top2)
-                        # Rgt
-                        # contour_temp.append(xy1)
-                        contour_temp.append(xy2)
-                        # Bot
-                        # contour_temp.append(bot1)
-                        # contour_temp.append(bot2)       
-                        
+                        r_t,r_b = (i,0),(i,y_black)
+                        contour_temp.append(r_t)
+                        contour_temp.append(r_b)
                         # Commit to big container
                         black_k_contours.append(contour_temp)
-
                         # Empty temp container
                         contour_temp = []
-                        cnt -= 1
-
+                        # Reset mode
+                        mode_flag = True
 
             white_k_count_Array = np.array(white_k_contours).reshape((-1,1,2)).astype(np.int32)
             white_k_count_Array = np.split(white_k_count_Array,52)
@@ -111,6 +71,10 @@ def Pre_comp_prep(crop_dim,bl_wh_lvls):     # Case by case comparison
             black_k_count_Array = np.array(black_k_contours).reshape((-1,1,2)).astype(np.int32)
             black_k_count_Array = np.split(black_k_count_Array,36)
             return white_k_count_Array, black_k_count_Array
+
+
+
+
 
         def midi_note_creation(white_k_count_Array,black_k_count_Array):
             # Midi Note Number constructor [clasified by white/black]
@@ -157,7 +121,6 @@ def Pre_comp_prep(crop_dim,bl_wh_lvls):     # Case by case comparison
             white_midi_notes = l
 
             # conform both in ascending order
-            # Compound_Note_Array = np.array()
             tem = []
             w_cnt = 0
             b_cnt = 0